+++ Programmation/Codes fonctionnels/Arduino/temp.cpp
@@ -10,7 +10,7 @@ bool Temp::init(){ // Initialisation du capteur
 }
 
 bool Temp::refresh(){  // Recuperation de la temperature
- double Temp;
+ float Temp;
  int RawADC = analogRead(9);  // Lecture de la tension
  Temp = log(((10240000/RawADC) - 10000));                                              // Conversion de la tension en temperature
  Temp = 1 / (0.001129148 + (0.000234125 + (0.0000000876741 * Temp * Temp ))* Temp );
